backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.1
python-dotenv==1.0.1
pymongo==4.6.1
razorpay==1.4.1
qrcode==7.4.2
python-multipart==0.0.9
pillow==10.2.0
pydantic==2.6.1 